UPSTREAM: ASoC: SOF: Intel: MTL: Don't access EM2
This reverts commit 2b5a30cafb2ef ("ASoC: SOF: Intel: MTL: Enable
DMI L1").

It came to our attention that the access to the EM2 register is restricted
to the DSP side on MTL compared to prior platforms.

Writing to it from the host side has no effect (negative or positive), it
is better to remove the code to not cause confusion and wrong impression.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
Link: https://lore.kernel.org/r/20230310133454.15362-1-peter.ujfalusi@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 75034eb5e4c267d4ab93085b5d6fa3dab2535b34)

BUG=b:277036374
TEST=Test Audio use cases

Change-Id: I7037ca59f1d9bcb27689a76835321e95a9b392b2
Signed-off-by: Debi sahoo <debix.prasad.sahoo@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4452948
Reviewed-by: Sean Paul <sean@poorly.run>
Commit-Queue: Terry Cheong <htcheong@chromium.org>
Reviewed-by: Terry Cheong <htcheong@chromium.org>
Tested-by: Terry Cheong <htcheong@chromium.org>