FROMGIT: drm/i915/guc: Clean up of register capture search
The comparison in the search for a matching register capture node was
not the most readable. It was also assuming that a zero GuC id means
invalid, which it does not. So remove one invalid term, one redundant
term and re-format to keep each term on a single line, and only one
term per line.

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Alan Previn <alan.previn.teres.alexis@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230311063714.570389-3-John.C.Harrison@Intel.com
(cherry picked from commit 53c4e64c6a0fdadc972217343f4439a96542f56e
 https://anongit.freedesktop.org/git/drm-intel.git drm-intel-next)

BUG=b:265114629
TEST=i915 driver loaded successfully

Change-Id: I7391ad25f7e5b0e5c4a073d90936072d475becf8
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4447999
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Tested-by: Natarajan Balamurugan <balamurugan.natarajan@intel.corp-partner.google.com>
Reviewed-by: Sean Paul <sean@poorly.run>