UPSTREAM: soundwire: intel: simplify sync_go sequence
In the existing code, the SHIM_SYNC::SYNC_GO bit is set, and the code
waits for it to return to zero.

That second wait part is just wrong: the SYNC_GO bit is *write-only* so
there's no way to know if it's cleared by hardware. The code works
because the value for a read-only bit is zero, but that's really just
luck.

Simplify the sequence to a plain read-modify-write.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Reviewed-by: Rander Wang <rander.wang@intel.com>
Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>
Link: https://lore.kernel.org/r/20230314015410.487311-6-yung-chuan.liao@linux.intel.com
Signed-off-by: Vinod Koul <vkoul@kernel.org>
(cherry picked from commit 9c49a4dd6c861bd2636094712d6d874e257787e2)

BUG=b:286158870
TEST=Test Audio use cases

Change-Id: I2d0a15c8f3bcffec8cfcb6441cad7e80612824c4
Signed-off-by: Subbarao Aketi <aketix.subbarao@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4600967
Reviewed-by: Terry Cheong <htcheong@chromium.org>
Commit-Queue: Terry Cheong <htcheong@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Terry Cheong <htcheong@chromium.org>