FROMGIT: dt-bindings: gpu: mali-bifrost: Split out MediaTek power-domains variation
In preparation for adding new bindings for new MediaTek SoCs, split out
the power-domains variation from the `else` in the current
mediatek,mt8183-mali conditional.

The sram-supply part is left in place to be disallowed for anything
that is not compatible with "mediatek,mt8183-mali" as this regulator
is MediaTek-specific and it is, and will ever be, used only for this
specific string due to the addition of the mediatek-regulator-coupler
driver.

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>
Tested-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: Rob Herring <robh@kernel.org>
Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230316102041.210269-2-angelogioacchino.delregno@collabora.com
(cherry picked from commit e82b84538ad8dddce4058a89aa3383d54f7935f8
 git://anongit.freedesktop.org/drm-misc drm-misc-next)

BUG=b:222013544
TEST=Boot to GUI on Jacuzzi

Change-Id: I4058a47b3a30ec87b4d0f583327924c3ced6a5ca
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4388478
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Fei Shao <fshao@chromium.org>