UPSTREAM: ASoC: SOF: ipc4/intel: Add missing mutex_unlock()
There was a missing mutex_unlock() in sof_ipc4_widget_free()
use_chain_dma if-branch that caused a static analysis error. The
branch should not be used in a normal working configuration and if its
used its an indication of a bad topology. Add missing mutex_unlock()
and a warning print if the if-branch is taken, and another warning
print to a symmetric place in sof_ipc4_widget_setup().

Fixes: ca5ce0caa67fa9 ("ASoC: SOF: ipc4/intel: Add support for chained DMA")
Reported-by: kernel test robot <lkp@intel.com>
Reported-by: Dan Carpenter <error27@gmail.com>
Link: https://lore.kernel.org/r/202303222050.dCw0fPCW-lkp@intel.com/
Signed-off-by: Jyri Sarha <jyri.sarha@linux.intel.com>
Link: https://lore.kernel.org/r/20230322181830.574635-1-jyri.sarha@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 935d31fdda2c69324b3eeb648a73fdedf4131474)

BUG=b:277036374
TEST=Test Audio use cases

Change-Id: Icc155c9b35107efd1d5c8904e819fe3de08e610d
Signed-off-by: Debi sahoo <debix.prasad.sahoo@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4453010
Commit-Queue: Terry Cheong <htcheong@chromium.org>
Tested-by: Terry Cheong <htcheong@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Terry Cheong <htcheong@chromium.org>