UPSTREAM: regmap: Add RAM backed register map
Add a register map that is a simple array of memory, for use in
KUnit testing of the framework. This is not exposed in regmap.h
since I can't think of a non-test use case, it is purely for use
internally. To facilitate testing we track if registers have been
read or written to.

Signed-off-by: Mark Brown <broonie@kernel.org>
Link: https://lore.kernel.org/r/20230324-regmap-kunit-v2-1-b208801dc2c8@kernel.org
(cherry picked from commit f6352424e37e7bf72291ceab87dc620172be0999)

BUG=b:286158870
TEST=Test Audio use cases

Change-Id: Ib996c65f9ec2dad60c4214027484aa28b328cf6c
Signed-off-by: Subbarao Aketi <aketix.subbarao@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4601124
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Terry Cheong <htcheong@chromium.org>
Reviewed-by: Terry Cheong <htcheong@chromium.org>
Commit-Queue: Terry Cheong <htcheong@chromium.org>