FIXUP: FROMLIST: thermal: mediatek: add another get_temp ops for thermal sensors
The original patch changed the thermal zone device's private data, i.e.
tz->devdata, adding one layer of indirection. It seems that during the
rebase, the change on the thermal zone registration side was preserved,
but on the temperature readout side it was not. This causes the
temperature readout to crash, as it tries to dereference a pointer of
an incorrect type.

Fix this by adding back the change in mtk_read_temp() and
mtk_read_sensor_temp().

BUG=b:236331727
TEST=Build and boot on kukui / jacuzzi
UPSTREAM-TASK=b:204226007, b:236331727

Change-Id: I95dc0686f7e773a8efbd840c30352fe7895f7037
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4384325
Reviewed-by: Fei Shao <fshao@chromium.org>