CHROMIUM: arm64: dts: mediatek: mt8188-geralt: Add panel AVEE ramp delay
According to the measurement, the AVEE supply rail from TPS65132
requires a few milliseconds to reach the target voltage, more
specifically the tRW delay should start from the second edge fall.

Add a 10ms enable ramp delay to the AVEE rail based on the measurement.
This assures that we don't violate the spec when powering on the panel.

BUG=b:311139883
BUG=b:273652518
TEST=conduct measurement with oscilloscope
UPSTREAM-TASK=b:244281402

Signed-off-by: Fei Shao <fshao@chromium.org>
Change-Id: I4dd9eea0b69e915b76b05eb29dc12ec4db0f5c1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4383869
Commit-Queue: Yu-Che Cheng <giver@chromium.org>
Reviewed-by: Chen-Yu Tsai <wenst@google.com>
Tested-by: Yu-Che Cheng <giver@chromium.org>