UPSTREAM: accel/habanalabs: ignore false positive razwi
In Gaudi2 asic, PSOC RAZWI may cause in HBW or LBW. The address that
caused the error is read from HW register and printed by the Driver.
There are cases where the Driver receives an indication on PSOC
RAZWI error but the address value is zero. In that case, the indication
is a false positive.
The Driver should not "count" a PSOC RAZWI event error when the
caused the address is zeroed.

Signed-off-by: Tal Cohen <talcohen@habana.ai>
Reviewed-by: Oded Gabbay <ogabbay@kernel.org>
Signed-off-by: Oded Gabbay <ogabbay@kernel.org>
(cherry picked from commit 1464fbd8bab958b771af3214bd5cb52cd5f612ed)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:307778453
TEST=cq dry run

Backported using forklift.py

Change-Id: Ib20317b3e2a799350d9d5bdc43fb45527671a2ee