UPSTREAM: tee: optee: Add SMC for loading OP-TEE image
Adds an SMC call that will pass an OP-TEE binary image to EL3 and
instruct it to load it as the BL32 payload. This works in conjunction
with a feature added to Trusted Firmware for ARMv8 and above
architectures that supports this.

The main purpose of this change is to facilitate updating the OP-TEE
component on devices via a rootfs change rather than having to do a
firmware update. Further details are linked to in the Kconfig file.

Signed-off-by: Jeffrey Kardatzke <jkardatzke@chromium.org>
Reviewed-by: Sumit Garg <sumit.garg@linaro.org>
Signed-off-by: Jeffrey Kardatzke <jkardatzke@google.com>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>
(cherry picked from commit f3040daab8104ac68e2b241bc93288f63b88a3c7)

BUG=b:246837563
TEST=CQ

Change-Id: Ibd09f38df626d822e94e27700ee8e06ba3a53788
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4765479
Reviewed-by: Sean Paul <sean@poorly.run>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Auto-Submit: Jeffrey Kardatzke <jkardatzke@google.com>